Beam
Beam，未完，还差一个大方法

+++ NVexFlow/NVexFlow/Model/StemmableNoteModel.cs
@@ -2,7 +2,7 @@
 {
     public class Glyph4StemmableNote:Glyph4Note
     {
-        public double beamCount;
+        public int beamCount;
     }
     
 }
